+++ src/niprefs/nodeutils.nim
@@ -1,13 +1,17 @@
-import std/[strformat, options, tables, macros]
+import std/[strformat, tables, macros]
 
 import toml_serialization
 import toml_serialization/value_ops
 
-export value_ops except len
+export value_ops except len, `[]`
 
 type
   TomlArray* = seq[TomlValueRef]
   TomlTables* = seq[TomlTableRef]
+
+template assertKind(node: TomlValueRef, kinds: varargs[TomlKind]) = 
+  assert not node.isNil
+  assert node.kind in kinds
 
 proc newTInt*(val: int64 = default int64): TomlValueRef = 
   TomlValueRef(kind: TomlKind.Int, intVal: val)
@@ -54,10 +58,10 @@ proc newTNode*(val: TomlDateTime): TomlValueRef =
 proc newTNode*(val: string): TomlValueRef = 
   newTString(val)
 
-proc newTNode*(val: TomlArray): TomlValueRef = 
+proc newTNode*(val: openArray[TomlValueRef]): TomlValueRef = 
   newTArray(val)
 
-proc newTNode*(val: TomlTables): TomlValueRef = 
+proc newTNode*(val: openArray[TomlTableRef]): TomlValueRef = 
   newTTables(val)
 
 proc newTNode*(val: TomlTableRef): TomlValueRef = 
@@ -69,35 +73,35 @@ proc newTNode*(val: openArray[(string, TomlValueRef)]): TomlValueRef =
 proc newTNode*(val: TomlValueRef): TomlValueRef = val
 
 proc getInt*(node: TomlValueRef): int64 = 
-  assert node.kind == TomlKind.Int
+  node.assertKind TomlKind.Int
   node.intVal
 
 proc getFloat*(node: TomlValueRef): float64 = 
-  assert node.kind == TomlKind.Float
+  node.assertKind TomlKind.Float
   node.floatVal
 
 proc getBool*(node: TomlValueRef): bool = 
-  assert node.kind == TomlKind.Bool
+  node.assertKind TomlKind.Bool
   node.boolVal
 
 proc getDateTime*(node: TomlValueRef): TomlDateTime = 
-  assert node.kind == TomlKind.DateTime
+  node.assertKind TomlKind.DateTime
   node.dateTime
 
 proc getString*(node: TomlValueRef): string = 
-  assert node.kind == TomlKind.String
+  node.assertKind TomlKind.String
   node.stringVal
 
 proc getArray*(node: TomlValueRef): TomlArray = 
-  assert node.kind == TomlKind.Array
+  node.assertKind TomlKind.Array
   node.arrayVal
 
 proc getTables*(node: TomlValueRef): TomlTables = 
-  assert node.kind == TomlKind.Tables
+  node.assertKind TomlKind.Tables
   node.tablesVal
 
 proc getTable*(node: TomlValueRef): TomlTableRef = 
-  assert node.kind == TomlKind.Table
+  node.assertKind TomlKind.Table
   node.tableVal
 
 macro toToml*(obj: untyped): TomlValueRef =
@@ -146,18 +150,48 @@ macro toToml*(obj: untyped): TomlValueRef =
   else:
     result = newCall("newTNode", obj)
 
-proc `[]=`*(obj: TomlValueRef, index: int, val: TomlValueRef) =
-  assert obj.kind == TomlKind.Array
+macro toTTables*(obj: untyped): TomlValueRef = 
+  obj.expectKind(nnkBracket)
+  if obj.len == 0: return newCall("newTTables")
+  result = newNimNode(nnkBracket)
 
-  obj.arrayVal[index] = val
+  for ele in obj:
+    result.add newCall("getTable", newCall("toToml", ele))
 
-proc `[]=`*[T: not TomlValueRef](obj: TomlValueRef, index: int, val: T) =
-  assert obj.kind == TomlKind.Array
+  result = newCall("newTTables", result)
 
-  obj.arrayVal[index] = val.newTNode()
+template toTTable*(body: untyped): TomlTableRef = 
+  toToml(body).getTable()
 
-proc `[]=`*[T: not TomlValueRef](obj: TomlValueRef, key: string, val: T) =
-  obj[key] = newTNode(val)
+proc `[]`*(node: TomlValueRef, name: string): TomlValueRef =
+  node.assertKind TomlKind.Table
+  node.tableVal[name]
+
+proc `[]`*(node: TomlValueRef, index: int): TomlValueRef =
+  node.assertKind TomlKind.Array, TomlKind.Tables
+
+  case node.kind
+  of TomlKind.Array:
+    result = node.arrayVal[index]
+  of TomlKind.Tables:
+    result = node.tablesVal[index].newTTable()
+  else: discard
+
+proc `[]=`*(node: TomlValueRef, index: int, val: TomlValueRef) =
+  node.assertkind TomlKind.Array
+
+  node.arrayVal[index] = val
+
+proc `[]=`*(node: TomlValueRef, index: int, val: TomlTableRef) =
+  node.assertkind TomlKind.Tables
+
+  node.tablesVal[index] = val
+
+proc `[]=`*[T: not TomlValueRef](node: TomlValueRef, index: int, val: T) =
+  node[index] = val.newTNode()
+
+proc `[]=`*[T: not TomlValueRef](node: TomlValueRef, key: string, val: T) =
+  node[key] = newTNode(val)
 
 proc `{}=`*[T: not TomlValueRef](node: TomlValueRef, keys: varargs[string], value: T) =
   node{keys} = newTNode(value)
@@ -208,12 +242,11 @@ proc `$`*(node: TomlValue): string =
     result = $node.tableVal
 
 proc `$`*(node: TomlValueRef): string = 
-  if node.isNil:
-    raise newException(ValueError, "Attempted to read from nil TomlValueRef node. You must initialize it first")
-  else:
-    $node[]
+  assert not node.isNil
+  $node[]
 
 proc `==`*[T: not TomlValueRef](node: TomlValueRef, val: T): bool = 
+  assert not node.isNil
   node == val.newTNode()
 
 proc tDateTime*(date: Option[TomlDate] = none(TomlDate), time: Option[TomlTime] = none(TomlTime), zone: Option[TomlTimeZone] = none(TomlTimeZone)): TomlDateTime = 
@@ -238,7 +271,7 @@ proc tTimeZone*(positiveShift = false, hourShift, minuteShift: int = 0): TomlTim
   TomlTimeZone(positiveShift: positiveShift, hourShift: hourShift, minuteShift: minuteShift)
 
 proc len*(node: TomlValueRef): int = 
-  assert node.kind in {TomlKind.Array, TomlKind.Tables, TomlKind.Table}
+  node.assertKind TomlKind.Array, TomlKind.Tables, TomlKind.Table
 
   case node.kind
   of TomlKind.Array:
@@ -250,22 +283,20 @@ proc len*(node: TomlValueRef): int =
   else: discard
 
 proc add*(node: TomlValueRef, table: TomlTableRef) = 
-  assert node.kind == TomlKind.Tables
+  node.assertKind TomlKind.Tables
 
   node.tablesVal.add(table)
 
 proc add*(node: TomlValueRef, val: TomlValueRef) = 
-  assert node.kind == TomlKind.Array
+  node.assertKind TomlKind.Array
 
   node.arrayVal.add(val)
 
 proc add*[T: not TomlValueRef](node: TomlValueRef, val: T) = 
-  assert node.kind == TomlKind.Array
-
-  node.arrayVal.add(val.newTNode())
+  node.add(val.newTNode())
 
 proc delete*(node: TomlValueRef, index: int) = 
-  assert node.kind in {TomlKind.Array, TomlKind.Tables}
+  node.assertKind TomlKind.Array, TomlKind.Tables
 
   case node.kind:
   of TomlKind.Array:
@@ -274,13 +305,17 @@ proc delete*(node: TomlValueRef, index: int) =
     node.tablesVal.delete(index)
   else: discard
 
+proc contains*(node: TomlValueRef, val: TomlTableRef): bool = 
+  node.assertKind TomlKind.Tables
+
+  val in node.tablesVal
+
 proc contains*[T: not TomlValueRef](node: TomlValueRef, val: T): bool =
-  assert node.kind == TomlKind.Array
   val.newTNode() in node
 
 proc hasKey*(node: TomlValueRef, keys: varargs[string]): bool = 
   ## Traverses the node and checks if the given key exists.
-  assert node.kind == TomlKind.Table
+  node.assertKind TomlKind.Table
 
   var table = node
   for key in keys:
@@ -292,7 +327,7 @@ proc hasKey*(node: TomlValueRef, keys: varargs[string]): bool =
   result = true
 
 iterator items*(node: TomlValueRef): TomlValueRef = 
-  assert node.kind in {TomlKind.Array, TomlKind.Tables}
+  node.assertKind TomlKind.Array, TomlKind.Tables
 
   case node.kind
   of TomlKind.Array:
@@ -302,7 +337,7 @@ iterator items*(node: TomlValueRef): TomlValueRef =
   else: discard
 
 iterator pairs*(node: TomlValueRef): (string, TomlValueRef) = 
-  assert node.kind == TomlKind.Table
+  node.assertKind TomlKind.Table
 
   for key, val in node.tableVal:
     yield (key, val)